+++ Changelog.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Changes
+- Add `new` method to ValueObject to easily update the object without unfreezing it
 - Include more information in the internal non-translated error messages. E.g. the max size of a LessThenEqual definition
 - Renamed GreaterThen, GreaterThenEqual, LessThen and LessThenEqual to fix typo (Then VS Than) Backwards compatibility is ensured